backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
click==8.1.8
fastapi==0.115.12
h11==0.16.0
idna==3.10
isort==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
ruff==0.11.8
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
